requirements.txt
google-genai>=1.0.0
mcp>=1.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
